Adapt RPC module to libftd
The whole RPC implementation could be heavily improved by proper trait/structure usage. But this is left for future PRs
