+++ tricore-docker/rpc-api/Cargo.toml
@@ -12,7 +12,10 @@ repository.workspace = true
 [dependencies]
 ciborium = "0.2.0"
 serde = { version = "1.0", features = ["derive"]}
-ftd-api = { path = "../ftd-api" }
 anyhow = "1.0.69"
 log = "0.4.0"
 tricore-common = { path = "../../tricore-common", features = ["serde"]}
+
+[dependencies.libftd2xx]
+version = "0.32.4"
+features = ["static"]
