+++ tricore-docker/rpc-api/src/lib.rs
@@ -2,5 +2,5 @@
 //! the docker and the rust application, once for the ftdi driver simulation
 //! and the other for communicating with the windows rust application.
 
-pub mod ftdi_commands;
+pub mod rpc;
 pub mod win_daemon;
